refactor:moved the rate lmiting funcitons to utils
