+++ apps/web/src/api/routes/v1_event.ts
@@ -6,25 +6,7 @@ import isbot from "isbot";
 import { EventService } from "server/services/EventService";
 import { RequestCache } from "server/services/RequestCache";
 import { RequestService } from "server/services/RequestService";
-
-import { RateLimiterRedis } from "rate-limiter-flexible";
-import { redis } from "server/db/redis";
-
-export const rateLimiter = new RateLimiterRedis({
-  storeClient: redis,
-  keyPrefix: "rateLimiter",
-  points: 10, // Number of points
-  duration: 10, // Per 10 seconds
-});
-
-export const checkRateLimit = async (ip: string) => {
-  try {
-    await rateLimiter.consume(ip);
-    return true;
-  } catch (rateLimiterRes) {
-    return false;
-  }
-};
+import { checkRateLimit } from "./utils";
 
 export function makeEventRoute() {
   const app = new Hono().post(
